+++ src/main/java/com/AppLink/app/AppApplication.java
@@ -1,13 +1,16 @@
-package com.applink.app;
+package com.AppLink.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.cache.annotation.EnableCaching;
 
 @SpringBootApplication
 @EnableCaching
-public class AppLauncher {
-    public static void main(String[] args) {
-        SpringApplication.run(AppLauncher.class, args);
-    }
+public class AppApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AppApplication.class, args);
+	}
+
 }
