--- src/main/java/com/AppLink/app/auth/dto/create/CreationController.java
@@ -1,4 +1,0 @@
-package com.AppLink.app.auth.dto.create;
-
-public class CreationController {
-}
