+++ src/main/java/com/AppLink/app/auth/dto/delete/Delete.java
@@ -1,0 +1,4 @@
+package com.AppLink.app.auth.dto.delete;
+
+public class Delete {
+}
