+++ src/main/java/com/AppLink/app/auth/dto/get/Get.java
@@ -1,0 +1,4 @@
+package com.AppLink.app.auth.dto.get;
+
+public class Get {
+}
