--- src/main/java/com/AppLink/app/auth/dto/links/LinksController.java
@@ -1,4 +1,0 @@
-package com.AppLink.app.auth.dto.links;
-
-public class LinksController {
-}
