+++ src/main/java/com/AppLink/app/controller/CreationController.java
@@ -1,0 +1,4 @@
+package com.AppLink.app.controller;
+
+public class CreationController {
+}
