+++ src/main/java/com/AppLink/app/controller/urlcontroller/UrlShortenerApplication.java
@@ -1,0 +1,4 @@
+package com.AppLink.app.controller.urlcontroller;
+
+public class UrlShortenerApplication {
+}
