+++ src/main/java/com/AppLink/app/entity/UrlEntity.java
@@ -1,0 +1,4 @@
+package com.AppLink.app.entity;
+
+public class UrlEntity {
+}
