+++ src/main/java/com/AppLink/app/security/Security.java
@@ -1,0 +1,4 @@
+package com.AppLink.app.security;
+
+public class Security {
+}
