--- src/main/java/com/AppLink/app/urlcontroller/UrlController.java
@@ -1,4 +1,0 @@
-package com.AppLink.app.urlcontroller;
-
-public class UrlController {
-}
