+++ src/main/java/com/AppLink/app/user/User.java
@@ -1,0 +1,4 @@
+package com.AppLink.app.user;
+
+public class User {
+}
