--- src/main/java/com/applink/app/configuration/ApplicationConfiguration.java
@@ -1,15 +1,0 @@
-package com.applink.app.configuration;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-@ComponentScan(value = "com.applink.app")
-public class ApplicationConfiguration {
-
-}
