--- src/main/java/com/applink/app/configuration/security/UserDetailsImp.java
@@ -1,21 +1,0 @@
-package com.applink.app.configuration.security;
-
-import com.applink.app.database.service.UserService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class UserDetailsImp implements UserDetailsService {
-    private final UserService userService;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) {
-        return userService.getUserByUsername(username);
-    }
-}
