--- src/main/java/com/applink/app/database/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.applink.app.database.repository;
-
-import com.applink.app.database.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    Optional<UserEntity> findByUsername(String username);
-}
