--- src/main/java/com/applink/app/database/service/UserService.java
@@ -1,18 +1,0 @@
-package com.applink.app.database.service;
-
-import com.applink.app.database.entity.UserEntity;
-import com.applink.app.database.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-    private final UserRepository userRepository;
-
-    public UserEntity getUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException(username));
-    }
-}
