--- src/main/java/com/applink/app/login/LoginController.java
@@ -1,7 +1,0 @@
-package com.applink.app.login;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController("loginController")
-public class LoginController {
-}
