--- src/main/java/com/applink/app/redirect/RedirectController.java
@@ -1,9 +1,0 @@
-package com.applink.app.redirect;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-
-@Controller("redirectController")
-@RequiredArgsConstructor
-public class RedirectController {
-}
