--- src/main/java/com/applink/app/registration/RegistrationController.java
@@ -1,7 +1,0 @@
-package com.applink.app.registration;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController("registrationController")
-public class RegistrationController {
-}
