--- src/main/java/com/applink/app/urls/UrlsController.java
@@ -1,7 +1,0 @@
-package com.applink.app.urls;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController("UrlsController")
-public class UrlsController {
-}
