--- src/main/resources/application-dev.properties
@@ -1,24 +1,0 @@
-#### HIBERNATE CONFIGURATION ###########
-spring.jpa.hibernate.ddl-auto=validate
-spring.jpa.show-sql=true
-
-#### DATASOURCE CONFIGURATION ##########
-spring.datasource.url=jdbc:h2:~/dev
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-
-
-#### FLYWAY CONFIGURATION ##############
-spring.flyway.locations=classpath:db/dev/migration
-spring.flyway.url=jdbc:h2:~/dev
-spring.flyway.user=sa
-spring.flyway.password=
-
-#### SWAGGER CONFIG ###########
-springdoc.api-docs.path=/api-docs
-springdoc.swagger-ui.operationsSorter=method
-springdoc.swagger-ui.path=/swagger-ui-custom.html
-
-#### GENERAL VARIABLES #########
-jjwt.token=defaultDevelopmentTokenForDevelopmentOnly1111
