--- src/main/resources/application-prod.properties
@@ -1,26 +1,0 @@
-#### HIBERNATE CONFIGURATION ########
-spring.jpa.hibernate.ddl-auto=validate
-
-
-#### DATASOURCE CONFIGURATION ########
-spring.datasource.url=${DATASOURCE_URL}
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.username=${DATASOURCE_USERNAME}
-spring.datasource.password=${DATASOURCE_PASSWORD}
-
-#### FLYWAY CONFIGURATION ###########
-spring.flyway.url=${DATASOURCE_URL}
-spring.flyway.user=${DATASOURCE_USERNAME}
-spring.flyway.password=${DATASOURCE_PASSWORD}
-spring.flyway.locations=classpath:db.prod.migration
-
-
-#### SWAGGER CONFIG ###########
-springdoc.api-docs.path=/api-docs
-springdoc.swagger-ui.path=/swagger-ui-custom.html
-springdoc.swagger-ui.operationsSorter=method
-
-
-
-#### GENERAL VARIABLES #########
-jjwt.token=${JWT_TOKEN}
