+++ src/main/resources/application.properties
@@ -1,6 +1,5 @@
 spring.application.name=App
-server.port=9999
-
-spring.profiles.active=${APP_PROFILE:dev}
-
-
+spring.datasource.driverClassName=org.postgresql.Driver
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
+spring.datasource.url=jdbc:postgresql://${DB_HOST}:${DB_PORT}/${DB_NAME}
