--- src/main/resources/db/dev/migration/V1_1__short_url_migration.sql
@@ -1,9 +1,0 @@
-CREATE TABLE IF NOT EXISTS urls(
-                                   id BIGINT PRIMARY KEY,
-                                   short_url VARCHAR(12) UNIQUE ,
-                                   long_url VARCHAR(256) NOT NULL ,
-                                   user_id INTEGER REFERENCES users(id),
-                                   visited BIGINT DEFAULT 0,
-                                   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                                   expired_at TIMESTAMP DEFAULT (SELECT DATEADD('DAY', 14, CURRENT_TIMESTAMP))
-);
