--- src/main/resources/db/dev/migration/V1_2__seq_migration.sql
@@ -1,2 +1,0 @@
-CREATE SEQUENCE USERS_SEQ START WITH 5 INCREMENT BY 1;
-CREATE SEQUENCE URLS_SEQ  START WITH 5 INCREMENT BY 1;
