--- src/test/resources/application-test.properties
@@ -1,14 +1,0 @@
-
-
-## Datasource test config
-spring.test.database.replace=none
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-spring.flyway.user=postgres
-spring.flyway.password=postgres
-
-## Hibernate test configuration
-#----disable lazy init for testing
-spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
-spring.jpa.show-sql=true
